Fixed infinite loop with tauri

+++ mangadex-api-schema/src/v5/forum_thread.rs
@@ -10,7 +10,7 @@ use crate::FromResponse;
 #[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct ForumThreadObject<A, T = RelationshipType> {
     pub type_: T,
-    pub id: usize,
+    pub id: u32,
     pub attributes: A,
 }
 
@@ -28,5 +28,5 @@ impl<A, T> FromResponse for ForumThreadObject<A, T> {
 #[cfg_attr(feature = "serialize", derive(serde::Serialize))]
 #[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct ForumThreadAttributes {
-    pub replies_count: usize,
+    pub replies_count: u32,
 }
